Small simplification in skipmap codegen.
When generating skipmaps, the parse function generator keeps track of a
"line_entries" variable that's mostly introduced to condense line wrapping in
some cases where the number of entries is small. This does make the generated
code a bit shorter but it's also keeping track of a lot of state that hinders
readability. It's been raised in a separate CL review that maybe all this
additional complexity is not worth it. We can deal with a few extra generated
lines in the codegen.

PiperOrigin-RevId: 720276314