Randomize upb's map ordering
We use the address of a variable as the seed for upb's hash function, and this
allows us to get some randomness from ASLR.

The randomness is not perfect, especially since it won't necessarily help when
ASLR is not enabled. However, it seems to be enough to prevent unit tests from
relying on a deterministic map ordering.

PiperOrigin-RevId: 721049667