Rust: add support for importing a .proto file from another crate
Cross-crate imports in Rust are a bit tricky, because without outside
information, protoc does not know how to refer to Rust entities that were built
in a separate protoc invocation. There is no inherent way to know the name of
the crate that a particular .proto file lives in. Fortunately there is a
`crate_mapping` option in the Rust code generator that lets you provide a file
specifying the name of the crate that each .proto file belongs to, but some
work is required to generate this mapping file correctly.

A related issue is that when we invoke protoc, we need to be able to tell it
where to look on the filesystem to resolve imports of files in other crates.
And finally, when we invoke the C compiler, likewise we need to tell it where
to look for the headers associated with dependencies.

This change aims to solve these problems by introducing a `Dependency` struct
that contains all the necessary metadata about a cross-crate dependency,
including the name of the crate and the canonical paths of its .proto files. To
use a cross-crate dependency with the `CodeGen` helper class, you can just do
something like
`codegen.dependency(proto_dep!(other_crate))`. The `CodeGen` helper will then
handle the details of generating a crate mapping file and passing the necessary
flags to protoc and the C compiler.

For now, this setup requires you to add a boilerplate `get_dependency()`
function alongside your generated protobuf code. However, the plan is that
protoc will eventually generated this on its own so that no extra boilerplate
is required.

PiperOrigin-RevId: 731796490