Remove protobuf v25 dep from WORKSPACE.bzlmod
We depend on protobuf v25 for the purpose of compatibility tests. Ideally we
would get v25 directly from BCR, but as far as I can tell there is no easy way
to get multiple versions of protobuf from BCR in the same bzlmod build. There
is
[multiple_version_override](https://bazel.build/rules/lib/globals/module#multiple_version_override),
but this seems to be geared toward resolving diamond dependency issues and not
so much toward deliberately testing against multiple versions.

This change therefore takes a different approach and uses
[archive_override](https://bazel.build/rules/lib/globals/module#archive_override).
I copied the necessary
[patches](https://github.com/bazelbuild/bazel-central-registry/tree/main/modules/protobuf/25.6/patches)
from BCR to get the code working with bzlmod, but skipped the `rules_ruby`
patch since that has already been applied on the commit we are using. I also
slightly tweaked the first and last patches to give the module the name we want
(`com_google_protobuf_v25`).

I noticed that the v25 commit we depend on is not actually v25.0, but is v25.6
with a few additional fixes. I therefore renamed `com_google_protobuf_v25.0` to
just `com_google_protobuf_v25` to avoid confusion.

PiperOrigin-RevId: 741650024