Move `previous_release` dependency out of `WORKSPACE.bzlmod` and into `MODULE.bazel`
While I was at it, I upgraded the previous release to 30.2, since that is now
the most recent release.

`PROTOBUF_PREVIOUS_RELEASE` is not currently getting updated, and I deleted it
because I don't see an obvious way of referring to it from `MODULE.bazel`.
Hopefully soon we can drop it from `WORKSPACE` and then the only place to
update will be `MODULE.bazel`.

PiperOrigin-RevId: 741690468