hpb: Introduce HPB_INTERNAL_BACKEND and backend/cpp/cpp.h
This abrogates HPB_BACKEND_UPB and HPB_BACKEND_CPP. The backend is now controlled by a string_flag, which defaults to upb.

Note that INTERNAL was also added -- clarifying that this macro is for internal use only.

This continues the multibackend work by stubbing the second backend in cpp.h.

PiperOrigin-RevId: 748648056