Added field number validation to MiniDescriptor parsing.
This fixes a fuzz bug which could cause the decoder to accept invalid wire format.

Technically we should be validating a max field number for extensions too, but it's tricky to do while allowing it for MessageSet (we'd have to check in `upb_ExtensionRegistry_Add()`), and realistically there isn't much risk because the wire format decoder never allows a tag to exceed `uint32_t` anyway.

PiperOrigin-RevId: 761721168